Added quicksort and adjusted unit tests.
Added quick sort and unit test for said function. As well, many unit tests were not using the proper sort function so that was changed to be correct. Lastly, a second implementation of quicksort was added. The 2nd I fully understand the code, but the 1st implementation needs to be reviewed as I can't firmly grasp how all the code works.